(DI-136) Update build.sh to work with existing cloned workspace
Prior to this the build.sh script would always attempt to clone the libral repo, this commit updates the script to first check if there is already a libral dir present at `/usr/src/libral` which will result in the clone being skipped.

It also clears any existing build directory before creating it.