(DI-136) Dockerfile for building Go binaries
This commit adds a Dockerfile that builds on and extends the existing
`Dockerfile-el6-build-static` based images to allow the user to build
using the `libralgo` binding.

It enables building of the example `listproviders.go` client and is
intended to provide an example container which you can create your own
bespoke build tool.